requirements.txt
praw==6.0.0
aiohttp==3.3.2
discord.py==1.0.0a0
